students/K3341/Mihailov_Yriy/laboratory_works/Lab_1/app/core/config.py
import os
from dotenv import load_dotenv

load_dotenv()

DB_URL = os.getenv("DB_URL", "postgresql://postgres:123@localhost/time_manager_db")
JWT_SECRET = os.getenv("JWT_SECRET", "jwt_secret")
JWT_ALGORITHM = os.getenv("JWT_ALGORITHM", "HS256")
ACCESS_TOKEN_EXPIRE_MINUTES = int(os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES", "60"))
