students/K3341/Mihailov_Yriy/laboratory_works/Lab_1/requirements.txt
fastapi[all]
sqlmodel
psycopg2-binary
alembic
python-dotenv
PyJWT
